--- back-end/src/parser.ts
@@ -1,35 +1,0 @@
-import  { Request, Response, Router } from "express"
-import { Readable } from "stream";
-import readline from "readline";
-
-function validatorCSV(req: Request, res: Response){
-    if(req.file?.mimetype !== 'text/csv'){
-        res.send({Erro:'Tipo de arquivo não suportado, utilize um arquivo CSV.'}).status(400)
-        return false
-    }
-    return true
-}
-
-export default {
-    async postArquivo(req: Request, res: Response) {
-        if(!validatorCSV(req, res)) return 
-        
-        const arrayNames = []
-        const b = req.file?.buffer
-        
-        const readableFile = new Readable()
-        readableFile.push(b)
-        readableFile.push(null)
-    
-        const nameLine = readline.createInterface({
-            input: readableFile
-        })
-    
-        for await(let line of nameLine) {
-            const species_names = line.split(',')             //faz a leitura somente da primeira coluna do arquivo
-            arrayNames.push(species_names[0])
-        }
-        
-        return res.json(arrayNames);
-    }
-}
