+++ back-end/src/routes.ts
@@ -1,16 +1,13 @@
-import  { response, Router } from "express"
-import multer from 'multer';
-import ParserController from "./parser";
+import { Router } from 'express'
+import multer from 'multer'
+import { Controller } from './Controllers/Controller'
 
-const router =  Router()
+const router = Router()
 
-const upload = multer();
+const upload = multer()
 
-router.post('/users', (request, response) => {
-    return response.status(201).send();
-})
+const controller = new Controller()
 
-router.post('/img', upload.single("filecsv"), ParserController.postArquivo)
+router.post('/busca-taxonomica', upload.single('filecsv'), controller.buscaFloraDoBrasil)
 
-export { router } 
- 
+export { router }
